requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
marshmallow
flask-migrate
gunicorn
#psycopg2
#psycopg2-binary
mysql-connector-python
flask_mysqldb